// src/generated/model/invoiceResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { InvoiceItemResponse } from './invoiceItemResponse.ts';

export interface InvoiceResponse {
  id?: string;
  invoiceNumber?: string;
  feeAccountId?: string;
  enrollmentId?: string;
  invoiceDate?: string;
  dueDate?: string;
  status?: string;
  subtotal?: number;
  totalAmount?: number;
  remarks?: string;
  items?: InvoiceItemResponse[];
  createdAt?: string;
}